requirements.txt
python-telegram-bot[job-queue]==21.6
yfinance>=0.2.40